+++ packages/webview_flutter/webview_flutter_android/lib/src/android_webview_controller.dart
@@ -1476,6 +1476,7 @@ class AndroidNavigationDelegate extends PlatformNavigationDelegate {
       NavigationRequest(
         url: url,
         isMainFrame: isForMainFrame,
+        navigationType: NavigationType.unknown,
       ),
     );