+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Support return NavigationType in NavigationRequest from WKWebview.
+
 ## 3.16.6
 
 * Bumps androidx.annotation:annotation from 1.7.1 to 1.8.1.